+++ build.sc
@@ -104,6 +104,38 @@ object text extends SwamModule with PublishModule {
 
 }
 
+
+object generator extends SwamModule with PublishModule {
+
+  def moduleDeps = Seq(core, runtime)
+
+  def publishVersion = swamVersion
+
+  def artifactName = "swam-generator"
+
+  def ivyDeps = Agg(ivy"com.github.pureconfig::pureconfig-enumeratum:$pureconfigVersion")
+
+  def pomSettings =
+    PomSettings(
+      description = "Swam code generator library",
+      organization = "org.gnieh",
+      url = swamUrl,
+      licenses = Seq(swamLicense),
+      versionControl = VersionControl.github("satabin", "swam"),
+      developers = Seq(swamDeveloper)
+    )
+
+  object test extends Tests with ScalafmtModule {
+    def ivyDeps =
+      Agg(ivy"com.lihaoyi::utest:0.7.1", ivy"com.github.pathikrit::better-files:3.8.0", ivy"com.lihaoyi::pprint:0.5.5")
+
+    def moduleDeps = Seq(generator, text, util.test)
+
+    def testFrameworks = Seq("swam.util.Framework")
+  }
+
+}
+
 object runtime extends SwamModule with PublishModule {
 
   def moduleDeps = Seq(core)
