+++ generator/test/src/generator/TemplateGeneratorTests.scala
@@ -1,25 +1,14 @@
 package swam
-package import_generator
+package generator
 
-import java.io.File
-import java.nio.ByteBuffer
 import java.nio.file.Paths
 import java.util.concurrent.Executors
 
-import cats.effect.{Async, Blocker, IO}
-import pureconfig.ConfigSource
-import swam.runtime.{Engine, formats, imports}
-import swam.runtime.config.EngineConfiguration
-import swam.runtime.formats.DefaultFormatters
-import swam.runtime.import_generator.ImportGenerator
-import swam.runtime.imports.{AsInstance, AsInterface, TCMap}
-import swam.syntax.Section.Imports
-import swam.text.Compiler
-import swam.validation.Validator
+import cats.effect.{Blocker, IO}
+import swam.runtime.Engine
 import utest.{TestSuite, Tests, test}
 
 import scala.concurrent.ExecutionContext
-import scala.reflect.runtime._
 
 /**
   * @author Javier Cabrera-Arteaga on 2020-03-06
